Adding the D challenge from David Eckardt to the Ad Mgmt repo.
Even though David (unintentionally) gave Dylan a task was that
impossible and ridiculous on numerous counts, 99% completion of
this task has been fulfilled, minus the job of dealing with scope
nested classes at compile time because as Gavin Norman has pointed
out this is quite impossible. So, in David's own words, it seems
I've discovered a compiler error in the process of writing my
first bit of D code. :)